backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
openai
aiohttp
